Modifies the 'Gatherings' controller to only allow the owner of the event to delete it.

+++ app/controllers/gatherings_controller.rb
@@ -45,7 +45,13 @@ class GatheringsController < ApplicationController
   # Destroy
   def destroy
     gathering = Gathering.find(params[:id])
-    gathering.destroy
+    user = current_user
+
+    if user.id == gathering.user_id
+      gathering.destroy
+    else
+      head :unauthorized
+    end
   end
 
   # Strong Params
